Update stack-adt.rst
Improve usage of stack vs ``Stack`` in accordance with Issue #798. 

See line 199: previously *stack*, now *``Stack``*. I am unsure of how italics and code blocks interact and whether them being like that will make the website explode. Regardless, I'm unsure of how this unique case of emphasis should be handled. 

Also lines 5, 17, and 314: stacks to ``Stack``s rather than to ``Stacks`` as the Class is still called Stack not Stacks, and there could be confusion if the s is inside of the code block, though there may be disagreement on this matter, which is why I point it out